# pyoffer-agent-docs/openapi.yaml
openapi: 3.1.0
info:
  title: PyOffer Agent API
  version: 0.1.0
  description: API draft for the PyOffer Agent MVP.
servers:
  - url: http://localhost:8000/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ResumeUploadResponse:
      type: object
      required: [resume_id, file_name, status]
      properties:
        resume_id:
          type: string
          format: uuid
        file_name:
          type: string
        status:
          type: string
    ResumeAnalyzeRequest:
      type: object
      properties:
        target_direction:
          type: string
          enum: [python_backend, agent_engineer, llm_application]
    ResumeAnalyzeResponse:
      type: object
      required: [detected_direction, skills, weaknesses, suggestions]
      properties:
        detected_direction:
          type: string
        skills:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        project_deep_dive_points:
          type: array
          items:
            type: string
...
paths:
...
  /projects/{project_id}/resumes:
    post:
      summary: Upload resume
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Uploaded resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeUploadResponse"
  /resumes/{resume_id}/parse:
    post:
      summary: Parse resume text
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Parsed resume
          content:
            application/json:
              schema:
                type: object
  /resumes/{resume_id}/analyze:
    post:
      summary: Analyze resume
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeAnalyzeRequest"
      responses:
        "200":
          description: Resume analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeAnalyzeResponse"
...